update Resource.test to TS
